Refactor: consolidate VU meters
